fix(storefront): persist mock buy prices as uint256 amounts
Plain-language summary:
- Store negotiation prices and agreed amounts as decimal text in SQLite so normal 18-decimal token amounts do not overflow SQLite's signed INTEGER range.
- Keep storefront callers working with Python integer raw base-unit amounts instead of routing prices through floats.
- Fund the baked dev Anvil buyer with enough raw MOCK units for the bundled dev inventory, chunking transfers to avoid the Alkahest wrapper's per-call integer limit.
- Move the storefront lockfile to the current storefront-client wheel version required by the test-env generation path.

Architectural summary:
Issue #106 exposed two separate assumptions in the dev/mock buy path. First, EVM payment amounts are uint256-domain raw base units, while SQLite INTEGER is signed 64-bit; storing 150 * 10**18 as an INTEGER is not valid. Second, the baked Anvil state funded the documented buyer wallet with only 1,000,000,000 raw MOCK units even though the bundled inventory advertises 18-decimal prices. The durable fix is to carry raw integer amounts end-to-end, serialize the SQLite columns as decimal text where SQLite lacks the needed range, and generate a dev state with enough token balance for the documented mock flow.

Change groups:
- Add SQLite amount serialization/deserialization helpers for non-negative raw token amounts.
- Change negotiation message, local-state, and agreed-price amount columns to TEXT and migrate legacy INTEGER-backed tables in place.
- Remove float conversions from initial-price extraction, negotiation history reconstruction, and settlement verification calls.
- Widen policy persistence annotations to allow integer/string amount carriers during storage boundaries.
- Increase baked dev MOCK funding to 1,000 * 10**18 raw units and transfer it in chunks.
- Bump storefront's arkhai-storefront-client lower bound and lock entry to the current locally built wheel expected by make build-anvil-state.
- Add regression coverage for uint256-sized negotiation amounts and legacy amount-column migration.
- Remove the completed SQLite overflow TODO entry.

Boundary decisions and deferrals:
- Do not lower bundled inventory prices or change MOCK token decimals as a workaround.
- Do not commit generated Anvil state, local wheels, kubeconfigs, Terraform state, or scratch artifacts.
- Do not implement the broader issue-110 orchestration work in this branch.
- Do not rename the public agreed_price field; it remains the base-unit integer amount despite the legacy field name.
- Do not change integration-tests or provisioning-service dependency bounds in this slice; their >=0.11.0 constraints can still accept newer storefront-client wheels when those projects are re-resolved.

Validation:
- Ran .venv/bin/python -m pytest tests/unit/test_negotiation_thread.py tests/unit/test_extract_initial_price.py tests/unit/test_settlement_jobs.py tests/unit/test_escrow_verification.py tests/unit/services/test_admin_settle_service.py tests/integration/test_negotiate_controller.py -q from storefront/.
- Ran make build-anvil-state from the repository root.
- Ran cd storefront && uv lock --check --find-links ../.dist.
- Ran git diff --check.

Result:
The dev/mock buy path can persist bundled 18-decimal inventory prices without SQLite overflow, and regenerated dev Anvil state gives the documented buyer enough raw MOCK to escrow those prices.

+++ storefront/src/market_storefront/utils/settlement_jobs.py
@@ -149,7 +149,7 @@ async def start_settlement_job(
     await verify_escrow_for_settlement(
         escrow_uid=escrow_uid,
         seller_wallet=settings.wallet.address or "",
-        agreed_price=float(thread["agreed_price"]),
+        agreed_price=int(thread["agreed_price"]),
         agreed_duration_seconds=provision.duration_seconds,
         listing=our_order_dict,
         alkahest_client=alkahest_client,
